Fixed SearchClick
Right, so, I fixed the SearchBatch Search Button problem we were having
(@OtherDevs, please test).
Also, re-updated .gitignore to ignore another .properties file that was
going to cause conflicts. The one that keeps changing the directories
and shit. Yeah, fuck that guy.
Also, made it so that when the SearchBatch UI comes it, it automatically
clicks Search for you so you don't need to click search every time. It
will auto search the default (all) criteria (also reformatted some code
that was out of alignment.)