New Exit Standard
Now when you click on the great big red X on ALL UI's, it functions as a
back button.
The only UI that doesn't have this is the Login UI as this will close
the program.

You could be balls deep into the program and just hit the big red X and
it'll take you all the way back to the login screen.

======BUG======
If you have multiple UI's open and you press that sexy-ass X button on
the MainUI, it'll 'log you out' but keep all the other windows open. May
need to mouse-trap this.